pyproject.toml
[project]
name = "icecat-harvester"
version = "0.1.0"
description = "Tools to harvest and parse Open Icecat product data"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"